Adds compatibility with PHP 7
* Tests configuration updated
* Null renamed into NullObject
* Null is now an empty class that extends NullObject

Please note that this commit breaks compatibility and requires to bump to
a major release
