+++ .travis.yml
@@ -5,10 +5,11 @@ php:
   - 5.4
   - 5.5
   - 5.6
+  - 7
   - hhvm
 
 before_script:
-    - composer --dev install
+    - composer install
 
 script:
     - vendor/bin/phpunit
