+++ lib/ASN1/Universal/Null.php
@@ -10,45 +10,9 @@
 
 namespace FG\ASN1\Universal;
 
-use FG\ASN1\Object;
-use FG\ASN1\Parsable;
-use FG\ASN1\Identifier;
-use FG\ASN1\Exception\ParserException;
 
-class Null extends Object implements Parsable
-{
-    public function getType()
-    {
-        return Identifier::NULL;
-    }
-
-    protected function calculateContentLength()
-    {
-        return 0;
-    }
-
-    protected function getEncodedValue()
-    {
-        return;
-    }
-
-    public function getContent()
-    {
-        return 'NULL';
-    }
-
-    public static function fromBinary(&$binaryData, &$offsetIndex = 0)
-    {
-        self::parseIdentifier($binaryData[$offsetIndex], Identifier::NULL, $offsetIndex++);
-        $contentLength = self::parseContentLength($binaryData, $offsetIndex);
-
-        if ($contentLength != 0) {
-            throw new ParserException("An ASN.1 Null should not have a length other than zero. Extracted length was {$contentLength}", $offsetIndex);
-        }
-
-        $parsedObject = new self();
-        $parsedObject->setContentLength(0);
-
-        return $parsedObject;
-    }
+/**
+ * @deprecated Deprecated. Use {@link NullObject} instead.
+ */
+class Null extends NullObject{
 }
